requirements.txt
django
psycopg2-binary

google-generativeai
djangorestframework
requests
django-cors-headers
